% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{KarstID}
\alias{KarstID}
\title{Launch the KarstID shiny application}
\usage{
KarstID(...)
}
\arguments{
\item{...}{No argument are needed to launch the application.}
}
\description{
An implementation of common analyses of karst spring hydrographs
through a Shiny application. It includes recession curves, statistical,
classified discharges and correlational and spectral analyses. The
application also allows performing a classification of the hydrological
functioning and comparing the results to a database of 78 karst systems.
}
